+++ cpp_src/core/ft/areaholder.cc
@@ -38,17 +38,20 @@ void AreaHolder::Commit() {
 		}
 	}
 }
-void AreaHolder::AddWord(int start_pos, int size, int filed) {
+bool AreaHolder::AddWord(int start_pos, int size, int filed) {
 	Area thr_area{start_pos, start_pos + size};
-	insertArea(thr_area, filed);
+	return insertArea(thr_area, filed);
 }
-void AreaHolder::insertArea(const Area &area, int field) {
+bool AreaHolder::insertArea(const Area &area, int field) {
 	commited_ = false;
 	if (areas.size() <= size_t(field)) areas.resize(field + 1);
 	if (areas[field].empty() || !areas[field].back().Concat(area)) {
-		if (areas[field].size() < kMaxAreasInResult) areas[field].push_back(area);
+		if (areas[field].size() >= kMaxAreasInResult) return false;
+		areas[field].push_back(area);
 	}
+	return true;
 }
+
 AreaVec AreaHolder::GetSnippet(int field, int front, int back, int total_size) {
 	if (!commited_) Commit();
 	auto va = areas[field];